# Perf/span deserializer