--- a/src/Nethermind/Nethermind.Synchronization.Test/Trie/HealingTrieStoreTests.cs
+++ b/src/Nethermind/Nethermind.Synchronization.Test/Trie/HealingTrieStoreTests.cs
@@ -11,6 +11,7 @@ using Nethermind.Core.Crypto;
 using Nethermind.Core.Extensions;
 using Nethermind.Core.Test;
 using Nethermind.Core.Test.Builders;
+using Nethermind.Core.Utils;
 using Nethermind.Logging;
 using Nethermind.Synchronization.Trie;
 using Nethermind.Trie;
@@ -28,7 +29,7 @@ public class HealingTrieStoreTests
         TestMemDb db = new();
         db[TestItem.KeccakA.Bytes] = new byte[] { 1, 2 };
         HealingTrieStore healingTrieStore = new(db, Nethermind.Trie.Pruning.No.Pruning, Persist.EveryBlock, LimboLogs.Instance);
-        healingTrieStore.LoadRlp(TestItem.KeccakA, ReadFlags.None);
+        healingTrieStore.LoadRlp<byte[]?, ToArraySpanDeserializer>(ToArraySpanDeserializer.Instance, TestItem.KeccakA, ReadFlags.None);
     }
 
     [Test]
@@ -44,7 +45,7 @@ public class HealingTrieStoreTests
             .Returns(successfullyRecovered ? Task.FromResult<byte[]?>(rlp) : Task.FromResult<byte[]?>(null));
 
         healingTrieStore.InitializeNetwork(recovery);
-        Action action = () => healingTrieStore.LoadRlp(key, ReadFlags.None);
+        Action action = () => healingTrieStore.LoadRlp<byte[]?, ToArraySpanDeserializer>(ToArraySpanDeserializer.Instance, key, ReadFlags.None);
         if (isMainThread && successfullyRecovered)
         {
             action.Should().NotThrow();