--- a/src/Nethermind/Nethermind.Trie/Pruning/NullTrieNodeResolver.cs
+++ b/src/Nethermind/Nethermind.Trie/Pruning/NullTrieNodeResolver.cs
@@ -13,7 +13,9 @@ namespace Nethermind.Trie.Pruning
         public static readonly NullTrieNodeResolver Instance = new();
 
         public TrieNode FindCachedOrUnknown(Hash256 hash) => new(NodeType.Unknown, hash);
-        public byte[]? LoadRlp(Hash256 hash, ReadFlags flags = ReadFlags.None) => null;
-        public byte[]? TryLoadRlp(Hash256 hash, ReadFlags flags = ReadFlags.None) => null;
+        public T LoadRlp<T, TDeserializer>(TDeserializer deserializer, Hash256 hash, ReadFlags flags = ReadFlags.None) where TDeserializer : ISpanDeserializer<T>
+        {
+            return deserializer.Deserialize(null);
+        }
     }
 }