… src/Nethermind/Nethermind.Trie/TrieNodeResolverWithReadFlags.cs
… src/Nethermind/Nethermind.Trie/TrieStoreWithReadFlags.cs