- As requested, a span deserializer at IKeyValueStore.
- Why? The GetSpan requires a corresponding correct DangerouslyReleaseSpan call which can be dangerous and cause the IKeyValueStore to be hard to wrap.
- This cause EOACompressingStore to be unable to implement GetSpan. 
- Also, GetSpan can't use `ReadAhead` flag as the span is part of the iterator, so it can't do DangerouslyReleaseSpan. The `ReadAhead` flag is vital for batched trie visitor and snap serving on halfpath.

## Before

| Method                          | Mean       | Error     | StdDev    | Gen0     | Gen1     | Gen2    | Allocated   |
|-------------------------------- |-----------:|----------:|----------:|---------:|---------:|--------:|------------:|
| Scenarios                       |   123.9 us |   2.47 us |   5.76 us |   8.7891 |   2.1973 |       - |   144.65 KB |
| InsertAndHash                   | 3,218.0 us |  64.09 us | 158.41 us | 218.7500 | 128.9063 |       - |  3593.14 KB |
| InsertAndCommit                 | 5,570.9 us | 129.28 us | 377.13 us | 406.2500 | 281.2500 | 78.1250 |   5594.4 KB |
| ReadWithFullTree                | 3,621.5 us |  71.95 us | 177.84 us | 640.6250 |        - |       - | 10516.34 KB |
| ReadWithUncommittedFullTree     |   471.6 us |   7.01 us |   5.86 us |  23.4375 |        - |       - |      384 KB |
| ReadWithMemoryTrieStore         | 1,874.2 us |  35.80 us |  46.55 us |  66.4063 |        - |       - |  1091.95 KB |
| ReadWithMemoryTrieStoreReadOnly | 3,766.3 us |  74.38 us | 166.36 us | 640.6250 |   3.9063 |       - | 10517.31 KB |
| ReadAndDeserialize              | 3,623.5 us |  67.21 us | 135.78 us | 640.6250 |   7.8125 |       - | 10519.89 KB |


## After

| Method                          | Mean       | Error     | StdDev    | Gen0     | Gen1     | Gen2    | Allocated  |
|-------------------------------- |-----------:|----------:|----------:|---------:|---------:|--------:|-----------:|
| Scenarios                       |   130.2 us |   2.59 us |   6.55 us |  10.7422 |   2.1973 |       - |  175.77 KB |
| InsertAndHash                   | 3,216.7 us |  63.85 us | 113.49 us | 218.7500 | 128.9063 |       - | 3593.14 KB |
| InsertAndCommit                 | 5,756.3 us | 139.09 us | 396.82 us | 406.2500 | 289.0625 | 78.1250 | 5566.94 KB |
| ReadWithFullTree                | 3,690.5 us |  72.90 us | 121.79 us | 605.4688 |        - |       - | 9941.44 KB |
| ReadWithUncommittedFullTree     |   476.8 us |   9.52 us |  16.92 us |  23.4375 |        - |       - |     384 KB |
| ReadWithMemoryTrieStore         | 1,862.7 us |  36.16 us |  45.74 us |  66.4063 |        - |       - | 1091.95 KB |
| ReadWithMemoryTrieStoreReadOnly | 3,738.6 us |  73.77 us | 119.12 us | 601.5625 |        - |       - | 9942.41 KB |
| ReadAndDeserialize              | 3,806.7 us |  90.85 us | 262.11 us | 605.4688 |   7.8125 |       - |    9945 KB |

Well... that is disappointing.

## Changes

- _List the changes_

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [ ] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [ ] Yes
- [ ] No

#### If yes, did you write tests?

- [ ] Yes
- [ ] No

#### Notes on testing

_Optional. Remove if not applicable._

## Documentation

#### Requires documentation update

- [ ] Yes
- [ ] No

_If yes, link the PR to the docs update or the issue with the details labeled `docs`. Remove if not applicable._

#### Requires explanation in Release Notes

- [ ] Yes
- [ ] No

_If yes, fill in the details here. Remove if not applicable._

## Remarks

_Optional. Remove if not applicable._