@LukaszRozmej 2024-02-22
This doesn't make sense when we need to have RLP byte array for later. This makes sense when we directly transform to domain object - Blocks/Receipts?