@asdacap 2024-02-22
I was thinking that the EOA compression could be significant. It is not apparently. Also, I was thinking that this could be significant when allocating all CappedArray via a pool and therefore can be recovered relatively efficient. Does not work as well. Have not try testing BatchedTrieVisitor which can use both buffer pool and allow copying straight from iterator's buffer.